disallow action switch in the middle of another action
When the user chooses an action, failover, failback, or relocation, it must complete all the way
to the end. We'll fail it if the attempts to switch to a different action while in the middle of another.
For example; if a failover is in progress and the user requests a failback, the operation will fail.