Make a pv mw to be a one time shot for pv delivery
Using a ManifestWork to deliver PV metadata to the ManagedCluster can cause undesirable results
as the ManifestWork constantly try to keep the state of the Manifests the same as what has been
applied in the ManagedCluster. This is fine and desirable for VRG, but not for PVs. Once the PVs
metadata have been applied, they will change at some point and the change must persist.
In this commit, we use the MW simply to make the initial delivery of the PV metadata, and once
that is applied, we immediately clear the Manifests array, and then delete the MW before we proceed
to the next step. This guarantees that we will delete the MW before the PVs are modified. We clear
the Manifests array first so that the PVs metadata that we just applied don't get deleted when we
delete the MW.