Add a configmap option for the user to provide the name of the csidriver
The user can provide the name of the csidriver entry in the RamenConfig map.
If the entry does not exist or empty, the default cephfs csidriver name for
readonly volumes will be used: "openshift-storage.cephfs.csi.ceph.com". When
a match is found, a new StorageClass is created, customized for readonly volumes,
and also a readonly flag accessMode is passed to VolSync.
Having this user provided csidriver entry in the RamenConfig will allow a user to
disable usage of readonly volumes when using the default csidriver above. The user
can provide a fake csidriver entry in the RamenConfi map in order to test
a non-readonly volumes using the same cephfs csidriver.