Update function to return last operation result and error
This commit modifies the utility function that creates the ManifestWork to return
an additional value indicating the last operation result alongside the error. The
result can be one of three values: created, updated, or none. This change is
needed to track whether the ManifestWork resource was newly created, updated, or
left unchanged.

Signed-off-by: Benamar Mekhissi <bmekhiss@ibm.com>